:pencil: [#2661] -- Clean up logic action stories
* Moved storybook decorators to higher level
* Refactored VariableSelection to use story decorator
* Fixed Action story by enabling the feature flag in context
* Let storybook 'Actions' handle the callbacks
* Line up controls/context for form steps/form variables to get working action
* Wire up useArgs to make component/story interactive
