+++ src/openforms/js/components/admin/form_design/variables/ServiceFetchConfigurationForm.stories.mdx
@@ -1,8 +1,8 @@
 import {ArgsTable, Canvas, Meta, Story} from '@storybook/addon-docs';
 import {useFormik} from 'formik';
 
+import {FormLogicDecorator} from '../story-decorators';
 import ServiceFetchConfigurationForm from './ServiceFetchConfigurationForm';
-import {FormLogicDecorator} from './story-decorators';
 
 <Meta
   title="Form design/Service Fetch/ServiceFetchConfigurationForm"
