+++ src/openforms/js/components/admin/form_design/variables/ServiceFetchConfigurationPicker.stories.mdx
@@ -1,7 +1,7 @@
 import {ArgsTable, Canvas, Meta, Story} from '@storybook/addon-docs';
 
+import {FormLogicDecorator} from '../story-decorators';
 import ServiceFetchConfigurationPicker from './ServiceFetchConfigurationPicker';
-import {FormLogicDecorator} from './story-decorators';
 
 <Meta
   title="Form design/Service Fetch/ServiceFetchConfigurationPicker"
