--- src/openforms/js/components/admin/form_design/variables/story-decorators.js
@@ -1,12 +1,0 @@
-import {FormLogicContext} from '../Context';
-
-export const FormLogicDecorator = (Story, {args}) => (
-  <FormLogicContext.Provider
-    value={{
-      services: args.availableServices || [],
-      serviceFetchConfigurations: args.serviceFetchConfigurations || [],
-    }}
-  >
-    <Story />
-  </FormLogicContext.Provider>
-);
