+++ src/openforms/js/components/admin/forms/VariableSelection.stories.mdx
@@ -1,19 +1,49 @@
 import {ArgsTable, Canvas, Meta, Story} from '@storybook/addon-docs';
+import {useArgs} from '@storybook/client-api';
 
 import {FormContext} from 'components/admin/form_design/Context';
+import {FormDecorator} from 'components/admin/form_design/story-decorators';
 
 import VariableSelection from './VariableSelection';
 
-<Meta title="Form design/VariableSelection" component={VariableSelection} />
+<Meta
+  title="Form design/VariableSelection"
+  component={VariableSelection}
+  decorators={[FormDecorator]}
+  argTypes={{
+    availableFormSteps: {table: {disable: true}},
+    availableStaticVariables: {table: {disable: true}},
+    availableFormVariables: {table: {disable: true}},
+  }}
+/>
 
-# Field
+# Variable selection dropdown
 
 Select widget for form and static variables
 
-export const Template = args => (
-  <FormContext.Provider
-    value={{
-      formSteps: [
+export const Template = ({name, includeStaticVariables, onChange, filter}) => {
+  const [{value}, updateArgs] = useArgs();
+  return (
+    <VariableSelection
+      name={name}
+      value={value}
+      includeStaticVariables={includeStaticVariables}
+      onChange={event => updateArgs({value: event.target.value})}
+      filter={filter}
+    />
+  );
+};
+
+<Canvas>
+  <Story
+    name="Default"
+    args={{
+      name: 'variableSelection',
+      value: 'key2',
+      includeStaticVariables: true,
+      onChange: () => null,
+      filter: () => true,
+      availableFormSteps: [
         {
           configuration: {display: 'form'},
           formDefinition: 'foo',
@@ -24,7 +54,7 @@ export const Template = args => (
           validationErrors: [],
         },
       ],
-      staticVariables: [
+      availableStaticVariables: [
         {
           form: 'foo',
           formDefinition: 'foo',
@@ -32,7 +62,7 @@ export const Template = args => (
           key: 'key1',
         },
       ],
-      formVariables: [
+      availableFormVariables: [
         {
           form: 'bar',
           formDefinition: 'bar',
@@ -42,28 +72,10 @@ export const Template = args => (
       ],
     }}
   >
-    <VariableSelection {...args} />
-  </FormContext.Provider>
-);
-
-## Presentation
-
-<Canvas>
-  <Story
-    name="Default"
-    args={{
-      name: 'VariableSelection',
-      value: 'foo',
-      onChange: ({...args}) => {
-        null;
-      },
-      filter: () => true,
-    }}
-  >
     {Template.bind({})}
   </Story>
 </Canvas>
 
 ### Props
 
-<ArgsTable story="Default" />
+<ArgsTable of={VariableSelection} />
